Add Writing a Compiler in Go
Add the code for as I work through "Writing a Compiler in Go".

We use `go mod` instead of `direnv`. We create the `src` directory,
where we will add our own code over the course of this book, with a copy
of the 00 code (i.e. the code at the end of Writing an Interpreter in
Go).