[WACIG Chapter 5] Support global lets in compiler
Our compiler now supports let statements in the global scope.
We do so by using a symbol table, which allows us to link variable names
with an index value. The index value is used to instruct the VM to
get/set variables from an array (similar to how we handle constants).