[WACIG Chapter 7] First step in compiling calls w/ arguments
Take the first step in compiling function calls w/ arguments. We make
each function call op specify how many arguments the function has. In
addition, we compile the args the function is being called with (i.e.
the function args should be on the stack when we call the function).