[WACIG Chapter 8] Refactor builtins to object
This commit moves builtin functions from the `evaluator` package to the
`object` package.

Previously, `builtins` lived in the `evaluator` package. Builtins living
in this location becomes problematic when we try to add builtins to our
compiler/vm. Our compiler/vm should not depend on the `evaluator`
package, as the `evaluator` is strictly used for interpreted monkey.
However, we also do not want to duplicate our builtin function
definitions.

The best solution is refactoring our builtin functions into the `object`
package, which both interpreted and compiled Monkey execution paths
already rely on.