[WACIG Chapter 9] Add support for compiling free variables
Add support for compiling free variables.

The strategy is the following... when compiling a function, if we are
looking up a variable and find it in an enclosing function (and its not
a global), then we record it as a free variable.

After we are done compiling the function, we know all the free
variables, and can place them all on the stack and include them in the
closure op code.