[WACIG Chapter 9] Add support for closures in VM
Add support for closures in the VM. When we compile a closure, we first
copy all of the free variables off of the stack and into a list of free
variables. We then store those free variables (with their values) in a
list attached to the closure. When we want to get a free variable, we
first get the current closure, and then fetch it from the attached free
variables.