[Build a Simple DB] Complete chapter 8
Complete chapter 8, which means we are now using the b-tree format. We
have defined the layout of a leaf node, and support inserting key/value
pairs into a single-node tree. All behavior is the same as before the
b-tree refactor (i.e. when we just had a big list of rows), except
because we only support a single node (and there's a one-to-one mapping
between a node and a page), we can now only have as many rows in our db
as can be stored in a single page. Outside of that change, the behavior
remains the same.

As a supporting refactor, we need
to extract all constants into a single file. Otherwise, we get
into weird dependency cycles when trying to define constants.

We also have to get rid of the `size_of_attribute` macro, as it causes a
dependency cycle w/ the definition of the `Row` type.