Practice creating debian packages
Practice creating a debian package for zig. Eventually, I will create a
separate repository for the debs I want to publish, but for now, I just
want a scratch pad to experiment in.

Note, I can't actually include the "raw" .deb, because the binary is
larger than the 100MB limit on Github.

Long-term, can fix using `git-lfs` (or alternatively hosting on sr.ht,
etc...) but for now, just include the `.deb.`